Adds datatype conversion to get_player_by_name
Corrects data type for games_subs in FieldPlayer model

+++ models/players.py
@@ -15,7 +15,7 @@ class PlayerBase:
     minutes_pct: float
     minutes_per_start: float
     games_complete: int
-    games_subs: float
+    games_subs: int
     minutes_per_sub: float
     unused_subs: int
     plus_minus: float
